+++ src/python/zensols/actioncli/__init__.py
@@ -1,4 +1,5 @@
 from zensols.actioncli.log import *
+from zensols.actioncli.persist_work import *
 from zensols.actioncli.executor import *
 from zensols.actioncli.config import *
 from zensols.actioncli.yaml_config import *